+++ src/sagas/issue.js
@@ -1,11 +1,34 @@
 import { call, put, takeLatest } from "redux-saga/effects";
-import { fetchIssueData, createIssue, updateIssue } from "../service/apiRequest";
+import {
+  fetchIssueData,
+  createIssue,
+  updateIssue,
+} from "../service/apiRequest";
 import {
   FETCH_ISSUE_REQUEST,
   requestIssueSuccess,
   CREATE_ISSUE_REQUEST,
   UPDATE_ISSUE_REQUEST,
 } from "../actions";
+import { toast } from "react-toastify";
+
+const fullfilled = (message) => {
+  toast.success(message, {
+    className: "toast-success",
+    hideProgressBar: true,
+    autoClose: false,
+    position: "top-center",
+  });
+};
+
+const failed = (message) => {
+  toast.error(message, {
+    className: "toast-error",
+    hideProgressBar: true,
+    autoClose: false,
+    position: "top-center",
+  });
+};
 
 function* fetchIssueList(action) {
   try {
@@ -13,7 +36,7 @@ function* fetchIssueList(action) {
     const issueData = yield call(fetchIssueData, action);
     yield put(requestIssueSuccess(issueData));
   } catch (e) {
-    console.log(e);
+    failed("一覧の取得に失敗しました");
   }
 }
 
@@ -21,8 +44,9 @@ function* createIssueItem(action) {
   try {
     yield call(createIssue, action.payload);
     yield put({ type: FETCH_ISSUE_REQUEST, payload: { direction: "asc" } });
+    yield fullfilled("issueを作成しました");
   } catch (e) {
-    console.log(e);
+    failed("作成に失敗しました");
   }
 }
 
@@ -30,8 +54,9 @@ function* updateIssueItem(action) {
   try {
     yield call(updateIssue, action.payload);
     yield put({ type: FETCH_ISSUE_REQUEST, payload: { direction: "asc" } });
+    yield fullfilled("更新に成功しました");
   } catch (e) {
-    console.log(e);
+    failed("更新に失敗しました");
   }
 }
